Fix long->in64_t

+++ sw_gdg/bp4_osd.pyx
@@ -194,7 +194,7 @@ cdef class bp4_osd:
 
         assert total_count == self.encoding_input_count_z
 
-    cpdef np.ndarray[np.int32_t, ndim=2] decode(self, input_vector_x, input_vector_z):
+    cpdef np.ndarray[np.int64_t, ndim=2] decode(self, input_vector_x, input_vector_z):
         cdef int input_length_x = input_vector_x.shape[0]
         cdef int vn
 
@@ -220,7 +220,7 @@ cdef class bp4_osd:
         
         return stackchar2numpy(self.bp_decoding_x, self.bp_decoding_z, self.n)
 
-    cpdef np.ndarray[np.int32_t, ndim=2] camel_decode(self, input_vector_x, input_vector_z):
+    cpdef np.ndarray[np.int64_t, ndim=2] camel_decode(self, input_vector_x, input_vector_z):
         cdef int input_length_x = input_vector_x.shape[0]
         cdef int vn
